MSP: Use custom deleter for path pointers
The deleter does not call delete, but instead marks the path as unused.
The memory manager can then pick it up and reuse it.

Again, there's a tricky part in execution - we need to make sure that
the memory manager destructor is called *after* our data sets are torn
down.